refactored packages to separate backend .java files from any controller files in accordance to SoC principles
