+++ src/main/java/group/backend/Booking.java
@@ -1,4 +1,4 @@
-package group;
+package group.backend;
 
 import java.util.*;
