+++ src/main/java/group/backend/CustomerManagementSystem.java
@@ -1,4 +1,4 @@
-package group;
+package group.backend;
 
 import java.util.*;
 import java.io.File;
